requirements.txt
absl-py==0.9.0
asgiref==3.2.7
asn1crypto==1.2.0
astor==0.8.1
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.13.0
chardet==3.0.4
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
Django==3.0.6
django-crispy-forms==1.9.1
elasticsearch==7.7.0
fake-useragent==0.1.11
gast==0.3.3
grpcio==1.28.1
h5py==2.10.0
humanfriendly==8.2
idna==2.9
importlib-metadata==1.6.0
jsonpickle==1.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
menuinst==1.4.16
mock==4.0.2
numpy==1.18.3
pandas==1.0.3
Pillow==7.1.1
protobuf==3.11.3
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==223
requests==2.23.0
requests-aws4auth==0.9
ruamel-yaml==0.15.46
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
tqdm==4.45.0
urllib3==1.25.9
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.1.0
